requirements.txt
langchain>=0.1.0
langchain-groq
streamlit
python-dotenv
pypdf
